+++ Plugins/Player/Player.cpp
@@ -8,6 +8,7 @@
 #include "API/CGameObject.hpp"
 #include "API/CNWSScriptVar.hpp"
 #include "API/CNWSScriptVarTable.hpp"
+#include "API/CServerExoAppInternal.hpp"
 #include "API/CExoArrayListTemplatedCNWSScriptVar.hpp"
 #include "API/CNWSCreature.hpp"
 #include "API/CNWSQuickbarButton.hpp"
@@ -17,6 +18,8 @@
 #include "API/CNWSItem.hpp"
 #include "API/CNWRules.hpp"
 #include "API/CNWSCreatureStats.hpp"
+#include "API/CNWSPlayerCharSheetGUI.hpp"
+#include "API/CNWSPlayerInventoryGUI.hpp"
 #include "API/CTwoDimArrays.hpp"
 #include "API/CNWSModule.hpp"
 #include "API/C2DA.hpp"
@@ -77,6 +80,8 @@ Player::Player(const Plugin::CreateParams& params)
     REGISTER(SetPlaceableUsable);
     REGISTER(SetRestDuration);
     REGISTER(ApplyInstantVisualEffectToObject);
+    REGISTER(UpdateCharacterSheet);
+    REGISTER(OpenInventory);
 
 #undef REGISTER
 
@@ -581,6 +586,52 @@ ArgumentStack Player::ApplyInstantVisualEffectToObject(ArgumentStack&& args)
                     0.0f);                        // fDuration
         }
     }
+    return stack;
+}
+
+ArgumentStack Player::UpdateCharacterSheet(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+    if (auto *pPlayer = player(args))
+    {
+        const auto charSheet = pPlayer->m_pCharSheetGUI;
+        uint32_t msg = charSheet->ComputeCharacterSheetUpdateRequired(pPlayer);
+        if (msg)
+        {
+            auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
+            if (pMessage)
+                pMessage->WriteGameObjUpdate_CharacterSheet(pPlayer, msg);
+        }
+    }
+    return stack;
+}
+
+ArgumentStack Player::OpenInventory(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+
+    if (auto *pPlayer = player(args))
+    {
+        auto oidTarget = Services::Events::ExtractArgument<Types::ObjectID>(args);
+          ASSERT_OR_THROW(oidTarget != Constants::OBJECT_INVALID);
+        auto open = !!Services::Events::ExtractArgument<int32_t>(args);
+
+        CNWSPlayerInventoryGUI *pInventory = pPlayer->m_oidNWSObject == oidTarget ? pPlayer->m_pInventoryGUI :
+                                                                                    pPlayer->m_pOtherInventoryGUI;
+
+        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
+        if (pMessage && pInventory)
+        {
+            pMessage->SendPlayerToServerGuiInventory_Status(pPlayer, open, oidTarget);
+            pInventory->SetOpen(open, 0/*bClientDirected*/);
+
+            if (open)
+            {
+                pInventory->SetOwner(oidTarget);
+            }
+        }
+    }
+
     return stack;
 }